requirements.txt
boto3==1.10.39
botocore==1.13.39
docutils==0.15.2
jmespath==0.9.4
numpy==1.17.4
opencv-contrib-python==4.1.0.25
python-dateutil==2.8.0
python-dotenv==0.10.3
s3transfer==0.2.1
six==1.13.0
urllib3==1.25.7
